server/requirements.txt
# attrs==25.3.0
# beautifulsoup4==4.13.4
# certifi==2025.6.15
# cffi==1.17.1
# charset-normalizer==3.1.0
# h11==0.14.0
# idna==3.7
# langchain
# langchain-chroma
# numpy==1.26.2
# outcome==1.3.0.post0
# packaging==24.0
# pandas
# psycopg2-binary==2.9.10
# pycparser==2.22
# PySocks==1.7.1
# python-dateutil
# python-dotenv
# pytz==2025.2
# requests
# rich~=12.5.1
# selenium
# six==1.17.0
# sniffio==1.3.1
# sortedcontainers==2.4.0
# soupsieve==2.7
# trio==0.30.0
# trio-websocket==0.12.2
# typing_extensions==4.12.2
# tzdata==2025.2
# urllib3~=2.2.2
# webdriver-manager==4.0.2
# websocket-client==1.8.0
# wsproto==1.2.0
# crawl4ai
# chromadb
# ggshield
# unicorn
# undetected-chromedriver
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# crawl4ai==0.2.77
# supabase==2.0.2
# python-multipart==0.0.6
# python-docx==1.1.0
# PyPDF2==3.0.1
# pandas==2.1.4
# python-dotenv==1.0.0
# pydantic==2.5.2
# httpx==0.25.2
# Core attributes and utilities
# Core attributes and utilities
attrs
six
typing_extensions
packaging

# Web scraping and parsing
beautifulsoup4
soupsieve
selenium
webdriver-manager
undetected-chromedriver
crawl4ai

# HTTP and networking
certifi
charset-normalizer
idna
h11
requests
urllib3
httpx
websocket-client
wsproto
PySocks

# Async programming
sniffio
trio
trio-websocket
outcome
sortedcontainers

# Web framework
fastapi
uvicorn[standard]
python-multipart

# Database and API
supabase
psycopg2-binary

# LangChain and AI
langchain
langchain-chroma
chromadb

# Data processing
pandas
numpy
python-dateutil
pytz
tzdata

# Document processing
python-docx
PyPDF2

# Configuration and environment
python-dotenv
pydantic

# Development and security
rich
ggshield

# System utilities
cffi
pycparser
unicorn